+++ Pages/Medications/Edit.cshtml.cs
@@ -1,0 +1,77 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.EntityFrameworkCore;
+using Roshta.Data;
+using Roshta.Models;
+
+namespace Roshta.Pages_Medications
+{
+    public class EditModel : PageModel
+    {
+        private readonly Roshta.Data.ApplicationDbContext _context;
+
+        public EditModel(Roshta.Data.ApplicationDbContext context)
+        {
+            _context = context;
+        }
+
+        [BindProperty]
+        public Medication Medication { get; set; } = default!;
+
+        public async Task<IActionResult> OnGetAsync(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var medication =  await _context.Medications.FirstOrDefaultAsync(m => m.Id == id);
+            if (medication == null)
+            {
+                return NotFound();
+            }
+            Medication = medication;
+            return Page();
+        }
+
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more information, see https://aka.ms/RazorPagesCRUD.
+        public async Task<IActionResult> OnPostAsync()
+        {
+            if (!ModelState.IsValid)
+            {
+                return Page();
+            }
+
+            _context.Attach(Medication).State = EntityState.Modified;
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!MedicationExists(Medication.Id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+
+            return RedirectToPage("./Index");
+        }
+
+        private bool MedicationExists(int id)
+        {
+            return _context.Medications.Any(e => e.Id == id);
+        }
+    }
+}